backend/requirements.txt
asgiref @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/asgiref_1699497449686/work
Bottleneck @ file:///C:/b/abs_f7un855idq/croot/bottleneck_1709069969633/work
cffi @ file:///C:/b/abs_924gv1kxzj/croot/cffi_1700254355075/work
cryptography @ file:///C:/b/abs_e8cnom_zw_/croot/cryptography_1702071486468/work
Django @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/django_1699555153998/work
django-filter @ file:///home/conda/feedstock_root/build_artifacts/django-filter_1701783126982/work
django-rest-framework-mongoengine==3.4.1
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1711135016139/work
djongo==1.3.6
dnspython @ file:///C:/b/abs_449_z2xz29/croot/dnspython_1703097012342/work
greenlet @ file:///C:/b/abs_a6c75ie0bc/croot/greenlet_1702060012174/work
idna @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/idna_1699473483982/work
lxml @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/lxml_1699484695217/work
Markdown @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/markdown_1699503092856/work
mkl-fft @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mkl_fft_1699473528480/work
mkl-random @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mkl_random_1699473588250/work
mkl-service==2.4.0
mongoengine==0.28.2
numexpr @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/numexpr_1699503421264/work
numpy @ file:///C:/b/abs_c1ywpu18ar/croot/numpy_and_numpy_base_1708638681471/work/dist/numpy-1.26.4-cp312-cp312-win_amd64.whl#sha256=becc06674317799ad0165a939a7613809d0bee9bd328a1e4308c57c39cacf08c
pandas @ file:///C:/b/abs_3awk0iw2ab/croot/pandas_1709590545218/work/dist/pandas-2.2.1-cp312-cp312-win_amd64.whl#sha256=8f9837b9f672189b7e2df7dcf64b91243a78b0fad6e1125220b33a39b5c9e598
psycopg2 @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/psycopg2_1707437392972/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pymongo @ file:///D:/bld/pymongo_1711625084977/work
PyMySQL @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pymysql_1699563042671/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pytz_1699475334900/work
setuptools==68.2.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
SQLAlchemy @ file:///C:/b/abs_876dxwqqu8/croot/sqlalchemy_1705089154696/work
sqlparse==0.2.4
typing_extensions @ file:///C:/b/abs_23uoivfoy8/croot/typing_extensions_1705619982210/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
wheel==0.41.2
